Fix linters issues and add readme changes

+++ src/entities/Person/AdditionalInfo/AdditionalInfo.cpp
@@ -2,7 +2,7 @@
 
 using namespace person;
 
-AdditionalInfo::AdditionalInfo(std::string IDNumber, std::string taxNumber, std::string phoneNumber, std::string email, uint8_t age) {
+AdditionalInfo::AdditionalInfo(std::string IDNumber, std::string taxNumber, std::string phoneNumber, std::string email, uint32_t age) {
     this->idNumber = std::move(IDNumber);
     this->taxNumber = std::move(taxNumber);
     this->phoneNumber = std::move(phoneNumber);
@@ -26,6 +26,6 @@ std::string AdditionalInfo::GetEmail() {
     return this->email;
 }
 
-uint8_t AdditionalInfo::GetAge() {
+uint32_t AdditionalInfo::GetAge() {
     return this->age;
 }
