+++ src/entities/Student/Scores/Scores.cpp
@@ -3,10 +3,10 @@
 
 using namespace student;
 
-Scores::Scores(std::vector<std::map<SubjectsEnum, uint8_t>> scores) {
+Scores::Scores(std::vector<std::map<SubjectsEnum, uint32_t>> scores) {
     this->scores = std::move(scores);
 }
 
-std::vector<std::map<SubjectsEnum, uint8_t>> Scores::GetScores() {
+std::vector<std::map<SubjectsEnum, uint32_t>> Scores::GetScores() {
     return this->scores;
 }
